Remove emails.compat module entirely
Move to_unicode, to_native, to_bytes, formataddr to emails.utils.
Replace all compat aliases with direct stdlib imports:
- string_types -> str
- text_type -> str
- is_callable -> callable()
- urlparse -> urllib.parse
- NativeStringIO/StringIO/BytesIO -> io
- OrderedDict -> collections

Delete emails/compat package and remove from setup.py.
All 74 tests pass.
