--- emails/compat/_urlparse.py
@@ -1,6 +1,0 @@
-# encoding: utf-8
-
-try:
-    from urlparse import *
-except ImportError:
-    from urllib.parse import *
